SpaceLiDAR.jl

SpaceLiDAR._download

SpaceLiDAR.angle

SpaceLiDAR.angle

SpaceLiDAR.classify

SpaceLiDAR.download_curl

SpaceLiDAR.gedi_info

SpaceLiDAR.granule_from_file

SpaceLiDAR.granules_from_folder

SpaceLiDAR.in_bbox

SpaceLiDAR.info

SpaceLiDAR.info

SpaceLiDAR.netrc!

SpaceLiDAR.s3!

SpaceLiDAR.splitline

SpaceLiDAR.track_power

SpaceLiDAR.granule_from_file — Method

Generate granule from .h5 file.

SpaceLiDAR.granules_from_folder — Method

Generate granules from folder filled with .h5 files.

SpaceLiDAR.in_bbox — Method

Filter with bbox.

SpaceLiDAR.netrc! — Method

Writes/updates netrc file for ICESat-2 and GEDI downloads.

SpaceLiDAR._download — Method

This is a method because it will segfault if precompiled.

SpaceLiDAR.angle — Function

Rough approximation of the track angle on a Euclidian lon/lat plot.

SpaceLiDAR.angle — Function

Rough approximation of the track angle on a Euclidian lon/lat plot.

SpaceLiDAR.classify — Function

Retrieve all points as classified as ground in ATL08.

SpaceLiDAR.download_curl — Method

Hacked version of Base.download which adds cookies and (optional) netrc parsing.

SpaceLiDAR.gedi_info — Method

Derive info based on file id.

The id is built up as follows, see section 2.4 in the user guide GEDI02A2019110014613O01991T049050200101.h5 or newer (V2) GEDI02A2019242104318O0404601T023430200302V002.h5

SpaceLiDAR.info — Method

Derive info based on file id.

The id is built up as follows, see 1.2.5 in the user guide ATL03[yyyymmdd][hhmmss][ttttccss][vvvrr].h5

SpaceLiDAR.info — Method

Derive info based on file id.

The id is built up as follows, see 1.2.5 in the user guide ATL03[yyyymmdd][hhmmss][ttttccss][vvvrr].h5

SpaceLiDAR.s3! — Function

Change default download locatio n to a S3 bucket.

SpaceLiDAR.splitline — Function

Split a linestring if the next point is further than distance. Not using Haversine here, as we want to split on meridians and such.

SpaceLiDAR.track_power — Method

Return whether track is a strong or weak beam. See Section 7.5 The Spacecraft Orientation Parameter of the ATL03 ATDB.